backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
pydub==0.25.1
